+++ x/ibc/transfer/module.go
@@ -1,0 +1,52 @@
+package transfer
+
+import (
+	"fmt"
+
+	"github.com/cosmos/cosmos-sdk/types/module"
+
+	ibctransfer "github.com/cosmos/ibc-go/v5/modules/apps/transfer"
+	ibctransferkeeper "github.com/cosmos/ibc-go/v5/modules/apps/transfer/keeper"
+	"github.com/cosmos/ibc-go/v5/modules/apps/transfer/types"
+	porttypes "github.com/cosmos/ibc-go/v5/modules/core/05-port/types"
+
+	"github.com/evmos/evmos/v10/x/ibc/transfer/keeper"
+)
+
+var (
+	_ module.AppModule      = AppModule{}
+	_ module.AppModuleBasic = AppModuleBasic{}
+	_ porttypes.IBCModule   = IBCModule{}
+)
+
+// AppModuleBasic embeds the IBC Transfer AppModuleBasic
+type AppModuleBasic struct {
+	*ibctransfer.AppModuleBasic
+}
+
+// AppModule represents the AppModule for this module
+type AppModule struct {
+	*ibctransfer.AppModule
+	keeper keeper.Keeper
+}
+
+// NewAppModule creates a new 20-transfer module
+func NewAppModule(k keeper.Keeper) AppModule {
+	am := ibctransfer.NewAppModule(*k.Keeper)
+	return AppModule{
+		AppModule: &am,
+		keeper:    k,
+	}
+}
+
+// RegisterServices registers module services.
+func (am AppModule) RegisterServices(cfg module.Configurator) {
+	// Override Transfer Msg Server
+	types.RegisterMsgServer(cfg.MsgServer(), am.keeper)
+	types.RegisterQueryServer(cfg.QueryServer(), am.keeper)
+
+	m := ibctransferkeeper.NewMigrator(*am.keeper.Keeper)
+	if err := cfg.RegisterMigration(types.ModuleName, 1, m.MigrateTraces); err != nil {
+		panic(fmt.Sprintf("failed to migrate transfer app from version 1 to 2: %v", err))
+	}
+}